Separation into cpp and hpp
