+++ cubrid/CMakeLists.txt
@@ -234,6 +234,7 @@ set(QUERY_SOURCES
   ${QUERY_DIR}/string_regex_std.cpp
   ${QUERY_DIR}/string_regex_re2.cpp
   ${QUERY_DIR}/vacuum.c
+  ${QUERY_DIR}/ovfp_threahold_monitor.cpp
   ${QUERY_DIR}/xasl_cache.c
   )
 set(QUERY_HEADERS
